flags.c: Avoid compiler warning when casting our OffsetOf value
Strictly, our OffsetOf calculation "could" return a number as large as the
pointer size on the system - so the compiler will complain when we squash
it directly into an int. Casting via a size_t will avoid the warning.

Our structure offsets are small, so this is safe - if somewhat ugly.

+++ src/flags.c
@@ -163,7 +163,12 @@ FlagLayerActive (int n)
   return current_layer == n;
 }
 
-#define OffsetOf(a,b) (int)(&(((a *)0)->b))
+/* The cast to (int) is ONLY valid because we know we are
+ * taking offsets on structures where the offset will fit
+ * in an integer variable. It silences compile warnings on
+ * 64bit machines.
+ */
+#define OffsetOf(a,b) (int)(size_t)(&(((a *)0)->b))
 
 HID_Flag flags_flag_list[] = {
   {"style", FlagCurrentStyle, 0},
